Rename 'errors' to 'exceptions'. Other fixes
